ServiceReview

+++ TrafficAccidentsAnalysisBackEnd/src/main/java/com/trafficaccidentsanalysis/backend/repository/MotoristRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.trafficaccidentsanalysis.backend.model.Motorist;
 
 
-public interface MotoristRepository extends JpaRepository<Motorist, Long> {
+public interface MotoristRepository extends JpaRepository<Motorist, Integer> {
 }
